Properly annotate task inputs
This commit fixes compatibility with Gradle 7: the `CoverallsTask` task
is using properties which don't have proper input annotations. In Gradle 7,
this becomes an error.

The `env` property has been marked as an input, while the `logger` and
`sourceReportFactoryMap` ones have been classified as internal (they
are not user-configurable and as far as the source report factory map
is concerned, it seems to be used only in the task action).